backend/requirements.txt
Flask
Flask-Cors
Flask-MySQLdb
hugchat
mysql
mysql-connector-python
mysqlclient
PyYAML